mydigitallife.info isn't affected by this
MDL shouldn't be affected by this since we didn't have the settings enabled that caused the problem. We've also received an email from Cloudflare to confirm this.

It should be made clear that the list of sites using Cloudflare doesn't reflect who was and wasn't affected. Of course you should change your passwords anyway, just to be safe.

Source (you need to login to see this that's why I copy & paste the content over here):
https://forums.mydigitallife.info/threads/73308-List-of-Sites-possibly-affected-by-Cloudflare-s-Cloudbleed-HTTPS-Traffic-Leak?p=1321465&viewfull=1#post1321465